revert "refactor(lotto): model 패키지 수정"
This reverts commit 63a1199fc4bc7a22715d04b714eecf41219315b3.
