+++ docs/README.md
@@ -63,8 +63,8 @@
 - [ ] MVC 도입하자. 클래스가 많으니 빠르게 찾기 힘들다.
     - [x] 일단 집어넣기
     - [x] view 수정
-    - [x] controller 수정
-    - [x] model 수정
+    - [ ] controller 수정
+    - [ ] model 수정
     - [ ] Model은 Controller와 View에 의존하지 않아야 한다.
     - [ ] View에는 Model에만 의존해야 하고 Controller에는 의존하면 안 된다
     - [ ] View가 Model로 부터 데이터를 받을 때는, 사용자마다 다르게 보여주어야 하는 데이터에 대해서만 받아야 한다.
@@ -72,7 +72,7 @@
     - [ ] View 가 Model로부터 데이터를 받을 때, 반드시 Controller에서 받아야 한다.
 
 
-- [ ]  이름을 통해 의도를 드러냈는가(무엇을 어떻게 하는지, 어떤 일을 하고 무엇을 리턴하는지)
+- [ ]  이름을 통해 의도를 드러냈는가
 - [ ]  기능 요구 사항, 프로그래밍 요구 사항, 과제 진행 요구 사항 항목 잘 지켰는가
 - [ ]  [README.md](http://README.md)
    - [ ]  해당 프로젝트가 어떤 프로젝트인지, 어떤 기능을 담고 있는지 나타냈는가
