+++ src/main/java/lotto/Controller/CalculateLotto.java
@@ -1,6 +1,6 @@
 package lotto.Controller;
 
-import static lotto.Model.MatchNums.*;
+import static lotto.MatchNums.*;
 
 import java.util.List;
 import java.util.Map;
