+++ src/main/java/lotto/MatchNums.java
@@ -1,4 +1,4 @@
-package lotto.Model;
+package lotto;
 
 import java.util.HashMap;
